Additional changes and testing

+++ src/Core/NetTaskTask.Domain/Daos/Main/Person.cs
@@ -8,7 +8,7 @@ namespace NetTestTask.Domain.Daos.Main
         public string FirstName { get; set; }
         public string LastName { get; set; }
 
-        public long? AddressId { get; set; }
+        public int? AddressId { get; set; }
         public virtual Address Address { get; set; }
     }
 }
